Add ScrollView to ParticipantsView and fix filename typo
